# Conflicts:
#	src/main/java/faang/school/accountservice/dto/AccountDto.java
#	src/main/java/faang/school/accountservice/service/AccountService.java
#	src/main/java/faang/school/accountservice/service/AccountServiceImpl.java

+++ src/main/java/faang/school/accountservice/dto/AccountDto.java
@@ -16,7 +16,7 @@ public record AccountDto(
         @Max(20)
         @NotNull
         String accountNumber,
-        Long balance,
+        Double balance,
         @NotNull
         Owner owner,
         @NotNull
